move all text fields to Components
