+++ packages/amplication-client/src/Components/TextField.scss
@@ -1,4 +1,4 @@
-@import "../../Style/parameters.css";
+@import "../Style/parameters.css";
 
 .text-field {
   margin-bottom: var(--form-field-spacing);
