+++ packages/amplication-client/src/Entities/EntityForm.tsx
@@ -2,9 +2,9 @@ import React, { useMemo } from "react";
 import { Formik, Form } from "formik";
 import { Button } from "@rmwc/button";
 import "@rmwc/button/styles";
-import NameField from "./fields/NameField";
-import { DisplayNameField } from "./fields/DisplayNameField";
-import { PluralDisplayNameField } from "./fields/PluralDisplayNameField";
+import NameField from "../Components/NameField";
+import { DisplayNameField } from "../Components/DisplayNameField";
+import { PluralDisplayNameField } from "../Components/PluralDisplayNameField";
 import { ToggleField } from "../Components/ToggleField";
 import omit from "lodash.omit";
