+++ packages/amplication-client/src/Pages/EntityPageForm.tsx
@@ -7,8 +7,8 @@ import { Button } from "@rmwc/button";
 import "@rmwc/button/styles";
 import { DrawerHeader, DrawerTitle, DrawerContent } from "@rmwc/drawer";
 import * as types from "../types";
-import NameField from "../Entities/fields/NameField";
-import { TextField } from "../Entities/fields/TextField";
+import NameField from "../Components/NameField";
+import { TextField } from "../Components/TextField";
 import { CheckboxField } from "../Components/CheckboxField";
 import { SelectField } from "../Components/SelectField";
 import { MultiStateToggle } from "../Components/MultiStateToggle";
